Cleaned up some code and tests

+++ telemetry/event_counters.go
@@ -150,7 +150,7 @@ func ClaimCounter(
 // RelayMiningDifficultyGauge sets a gauge which tracks the integer representation
 // of the relay mining difficulty. The serviceId is used as a label to be able to
 // track the difficulty for each service.
-func RelayMiningDifficultyGauge(difficulty uint64, serviceId string) {
+func RelayMiningDifficultyGauge(difficulty float32, serviceId string) {
 	labels := []metrics.Label{
 		{Name: "type", Value: "relay_mining_difficulty"},
 		{Name: "service_id", Value: serviceId},
@@ -158,7 +158,7 @@ func RelayMiningDifficultyGauge(difficulty uint64, serviceId string) {
 
 	telemetry.SetGaugeWithLabels(
 		[]string{eventTypeMetricKeyGauge},
-		float32(difficulty),
+		difficulty,
 		labels,
 	)
 }
